Fixup all line endings to be Unix style
Files in this repo had mixed line endings CRLF (Windows) vs LF (Linux),
this was creating a headache when working with editors that do automatic
recognition of line ending styles e.g. vim.

All line endings are converted to LF(Linux) and additionally
.gitattributes is set to '* text=auto' this will ensure that git on
windows will automatically convert LF to CRLF on checkout and CRLF to LF
when added to the repository.

Note the exclusion on PNG files for this conversion.